+++ merlin/celery.py
@@ -49,6 +49,7 @@ from merlin.config import (
 )
 from merlin.log_formatter import FORMATS
 from merlin.router import route_for_task
+from merlin.config.configfile import CONFIG
 
 
 LOG = logging.getLogger(__name__)
@@ -100,10 +101,10 @@ app.conf.update(
     redis_max_connections=100000,
 )
 
-# Set a 24-hour timeout to acknowledge a task before it's available to grab
-# again.
+# Set a timeout to acknowledge a task before it's available to grab
+# again (default 24 hours).
 app.conf.broker_transport_options = {
-    "visibility_timeout": 86400,
+    "visibility_timeout": CONFIG.celery.visibility_timeout,
     "max_connections": 100,
 }