+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   
 ### Changed
 - The default rabbitmq vhost is now <user> instead of /<user>.
-- Changed default task timeout from 2 hours to 24 hours.
+- Changed default visibility timeout from 2 hours to 24 hours. Exposed this in the config
+  file.
 - The monitor function will now check the queues to determine when
   to exit.